Rename `http_client_util` to `scrub_secrets`.
Bug: b/422877302
