+++ base/cvd/cuttlefish/host/libs/web/http_client/url_escape.cc
@@ -18,6 +18,7 @@
 #include <curl/curl.h>
 
 #include <string>
+#include <string_view>
 
 namespace cuttlefish {
